+++ ui/src/history/history.page.tsx
@@ -1,35 +1,55 @@
-import { useEffect, useState } from "react";
-import { useServerObject } from "@src/shared/http-request";
 import { PagedModel, Trigger } from "@src/server-api";
-import { Col, Row, Stack } from "react-bootstrap";
+import { useServerObject } from "@src/shared/http-request";
+import useAutoRefresh from "@src/shared/use-auto-refresh";
 import HttpErrorView from "@src/shared/view/http-error.view";
 import PageView from "@src/shared/view/page.view";
 import ReloadButton from "@src/shared/view/reload-button.view";
 import TriggerItemView from "@src/shared/view/trigger-list-item.view";
+import TaskSelect from "@src/task/view/task-select.view";
+import { useState } from "react";
+import { Col, Form, Row, Stack } from "react-bootstrap";
 
 const HistoryPage = () => {
     const [page, setPage] = useState(0);
+    const [taskName, setTaskName] = useState("");
+    const [id, setId] = useState("");
+
     const triggers = useServerObject<PagedModel<Trigger>>(
         "/spring-tasks-api/history"
     );
 
     const doReload = () => {
-        triggers.doGet("?size=10&page=" + page);
+        triggers.doGet(
+            "?size=10&page=" + page + "&taskName=" + taskName + "&id=" + id
+        );
     };
 
-    useEffect(doReload, [page]);
-    useEffect(() => {
-        const intervalId = setInterval(doReload, 10000);
-        return () => clearInterval(intervalId);
-    }, [page]);
+    useAutoRefresh(10000, doReload, [page, taskName, id]);
 
     return (
         <>
             <Stack gap={1}>
                 <HttpErrorView error={triggers.error} />
-                <Row className="align-items-center">
-                    <div></div>
+                <Row>
                     <Col>
+                        <Form.Control
+                            type="text"
+                            placeholder="Search..."
+                            onKeyUp={(e) =>
+                                e.key == "Enter"
+                                    ? setId(
+                                          (e.target as HTMLInputElement).value
+                                      )
+                                    : null
+                            }
+                        />
+                    </Col>
+                </Row>
+                <Row className="align-items-center mb-2">
+                    <Col>
+                        <TaskSelect onTaskChange={setTaskName} />
+                    </Col>
+                    <Col className="align-items-center">
                         <PageView
                             onPage={(p) => setPage(p)}
                             data={triggers.data}